pyproject.toml
[project]
name = "agent-teams"
version = "0.1.0"
description = "数字员工协作平台 POC - 多Agent团队协作引擎"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "openai>=1.60.0",
    "tiktoken>=0.8.0",
    "aiosqlite>=0.20.0",
    "jinja2>=3.1.0",
    "markdown>=3.7",
    "ddgs>=9.0",
    "beautifulsoup4>=4.12.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
